requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
crypto==1.4.1
cycler==0.12.1
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.53.0
frozenlist==1.4.1
gcloud==0.18.3
googleapis-common-protos==1.63.0
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
multidict==6.0.5
Naked==0.1.32
numpy==2.0.0
oauth2client==4.1.3
packaging==24.1
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shellescape==3.8.1
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
